# backend/openapi.yaml
openapi: 3.1.0
info:
  title: AI E-Commerce Backend
  version: 0.1.0
paths:
...
  /api/v1/users/me:
    get:
      tags:
      - buyer
      summary: Get Me
      operationId: get_me_api_v1_users_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
    put:
      tags:
      - buyer
      summary: Update Me
      operationId: update_me_api_v1_users_me_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
...
  /api/v1/products:
    get:
      tags:
      - buyer
      summary: List Products
      operationId: list_products_api_v1_products_get
      parameters:
      - name: keyword
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Keyword
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      - name: search_type
        in: query
        required: false
        schema:
          type: string
          default: keyword
          title: Search Type
      - name: category_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Category Id
      - name: shop_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Shop Id
      - name: min_price
        in: query
        required: false
        schema:
          anyOf:
          - type: number
          - type: 'null'
          title: Min Price
      - name: max_price
        in: query
        required: false
        schema:
          anyOf:
          - type: number
          - type: 'null'
          title: Max Price
      - name: sort
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Sort
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/products/intent-search:
    post:
      tags:
      - buyer
      summary: Intent Search
      operationId: intent_search_api_v1_products_intent_search_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntentSearchIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/products/recommend:
    get:
      tags:
      - buyer
      summary: Recommend Products
      operationId: recommend_products_api_v1_products_recommend_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/products/{id}/reviews:
    get:
      tags:
      - buyer
      summary: List Product Reviews
      operationId: list_product_reviews_api_v1_products__id__reviews_get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/shops/{id}:
    get:
      tags:
      - buyer
      summary: Get Shop Public
      operationId: get_shop_public_api_v1_shops__id__get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/seller-requests/me:
    get:
      tags:
      - buyer
      summary: My Seller Request
      operationId: my_seller_request_api_v1_seller_requests_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
  /api/v1/seller-requests:
    post:
      tags:
      - buyer
      summary: Create Seller Request
      operationId: create_seller_request_api_v1_seller_requests_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SellerRequestIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/reports:
    post:
      tags:
      - buyer
      summary: Create Report
      operationId: create_report_api_v1_reports_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/notifications:
    get:
      tags:
      - buyer
      summary: List Notifications
      operationId: list_notifications_api_v1_notifications_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
  /api/v1/notifications/{id}/read:
    patch:
      tags:
      - buyer
      summary: Mark Notification Read
      operationId: mark_notification_read_api_v1_notifications__id__read_patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationReadIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/notifications/read-all:
    patch:
      tags:
      - buyer
      summary: Mark All Notifications Read
      operationId: mark_all_notifications_read_api_v1_notifications_read_all_patch
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
  /api/v1/chat/conversations:
    get:
      tags:
      - buyer
      summary: List Conversations
      operationId: list_conversations_api_v1_chat_conversations_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - buyer
      summary: Create Conversation
      operationId: create_conversation_api_v1_chat_conversations_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/chat/conversations/{conversation_id}/messages:
    get:
      tags:
      - buyer
      summary: List Messages
      operationId: list_messages_api_v1_chat_conversations__conversation_id__messages_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: integer
          title: Conversation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - buyer
      summary: Send Message
      operationId: send_message_api_v1_chat_conversations__conversation_id__messages_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: integer
          title: Conversation Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/chat/conversations/{conversation_id}/read:
    patch:
      tags:
      - buyer
      summary: Read Conversation
      operationId: read_conversation_api_v1_chat_conversations__conversation_id__read_patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: integer
          title: Conversation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/products/{id}:
    get:
      tags:
      - buyer
      summary: Get Product
      operationId: get_product_api_v1_products__id__get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/cart:
    get:
      tags:
      - buyer
      summary: Get Cart
      operationId: get_cart_api_v1_cart_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
  /api/v1/cart/items:
    post:
      tags:
      - buyer
      summary: Add Cart
      operationId: add_cart_api_v1_cart_items_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartAddIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/cart/items/{item_id}:
    put:
      tags:
      - buyer
      summary: Update Cart
      operationId: update_cart_api_v1_cart_items__item_id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: integer
          title: Item Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartUpdateIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - buyer
      summary: Delete Cart
      operationId: delete_cart_api_v1_cart_items__item_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: integer
          title: Item Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/orders:
    post:
      tags:
      - buyer
      summary: Create Order
      operationId: create_order_api_v1_orders_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/orders/my:
    get:
      tags:
      - buyer
      summary: List My Orders
      operationId: list_my_orders_api_v1_orders_my_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: from_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: From Date
      - name: to_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: To Date
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/orders/{id}:
    get:
      tags:
      - buyer
      summary: Get Order
      operationId: get_order_api_v1_orders__id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/orders/{id}/cancel:
    post:
      tags:
      - buyer
      summary: Cancel Order
      operationId: cancel_order_api_v1_orders__id__cancel_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/payments/{order_id}/pay:
    post:
      tags:
      - buyer
      summary: Pay Order
      operationId: pay_order_api_v1_payments__order_id__pay_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: integer
          title: Order Id
      - name: x-idempotency-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Idempotency-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/payments/{order_id}:
    get:
      tags:
      - buyer
      summary: Get Payment
      operationId: get_payment_api_v1_payments__order_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: integer
          title: Order Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/reviews:
    post:
      tags:
      - buyer
      summary: Create Review
      operationId: create_review_api_v1_reviews_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
...
components:
  schemas:
...
    ConversationIn:
      properties:
        seller_id:
          type: integer
          title: Seller Id
      type: object
      required:
      - seller_id
      title: ConversationIn
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    IntentSearchIn:
      properties:
        query:
          type: string
          title: Query
        page:
          type: integer
          title: Page
          default: 1
        page_size:
          type: integer
          title: Page Size
          default: 20
      type: object
      required:
      - query
      title: IntentSearchIn
...
    MessageIn:
      properties:
        content:
          type: string
          title: Content
      type: object
      required:
      - content
      title: MessageIn
    NotificationReadIn:
      properties:
        is_read:
          type: boolean
          title: Is Read
          default: true
      type: object
      title: NotificationReadIn
...
    ReportIn:
      properties:
        target_type:
          type: string
          title: Target Type
        target_id:
          type: integer
          title: Target Id
        reason:
          type: string
          title: Reason
      type: object
      required:
      - target_type
      - target_id
      - reason
      title: ReportIn
...
    SellerRequestIn:
      properties:
        shop_name:
          type: string
          title: Shop Name
        description:
          type: string
          title: Description
        contact:
          type: string
          title: Contact
      type: object
      required:
      - shop_name
      - description
      - contact
      title: SellerRequestIn
...
    UpdateMeIn:
      properties:
        full_name:
          type: string
          title: Full Name
      type: object
      required:
      - full_name
      title: UpdateMeIn
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login